refactor: restore color on ex
